+++ NEWS.md
@@ -5,6 +5,7 @@
 
 * added function `compound_table`
 * In forest plot functions, added formatting to display extreme values to ">999.9"." 
+* added `t_ae_ctc` adverse events table by highest NCI CTCAE grade function.
 
 # tern 0.5.0